feat(about): unify About with in-app dialog and fix Window menu build
- fix: SubmenuBuilder has no zoom() helper in Tauri 2.10.3, so use
  PredefinedMenuItem::maximize() for the Window menu's Zoom entry. muda
  maps the Maximize predefined type to native macOS window Zoom behavior,
  and the crate now compiles.
- feat: the macOS menu-bar About entry now opens the in-app About dialog
  instead of the native About sheet. The native sheet can only render
  name/version/copyright/icon, so a custom menu item (id "about") now
  emits a "menu:about" event (lib.rs on_menu_event) that opens
  AboutDialogContent via MenuAboutDialog, showing the same environment
  info (OS/arch/Tauri/repo) as Settings -> About. AboutDialogContent is
  factored out of AboutDialog so both surfaces share one body.
