+++ src/features/about/index.ts
@@ -1,3 +1,4 @@
 export { getAppInfo } from './api/aboutApi'
 export type { AppInfo } from './types'
 export { AboutDialog } from './ui/AboutDialog'
+export { MenuAboutDialog } from './ui/MenuAboutDialog'
